requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
idna==3.7
Js2Py==0.74
packaging==24.1
pipwin==0.5.2
psycopg2-binary==2.9.9
pyjsparser==2.7.1
PyJWT==2.8.0
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
